rework index page to show created and saved playlists
