+++ src/components/Index/ShowPlaylists/ShowPlaylists.jsx
@@ -1,0 +1,7 @@
+export default function ShowPlaylists() {
+    return (
+        <div>
+            <h1>Show Playlists</h1>
+        </div>
+    )
+}
